requirements.txt
autopep8==1.5.7
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
fastapi==0.63.0
h11==0.11.0
mysql-connector-python==8.0.24
protobuf==3.15.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.7.3
PyJWT==2.0.0
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
uvicorn==0.13.2
